Teach the synth some more chords, make chords configurable, add some chord progressions.
This change adds UI to specify the list of chords to play, and also allows choosing a key + chord progression to get chords for common progressions. I changed the strum pattern to play a single measure of each chord to make this work better.

I imported jtab to get a comprehensive list of guitar chords. The chord progressions come from teoria-chord-progression, so I used browserify to pull in teoria + that module and produce progressions.js usable in the browser.